Deduplicate shuffle and centralize currentTurnState helpers
Extract the three identical Fisher-Yates shuffle() copies into a single
shared module src/lib/game/shuffle.ts and import it in secret-villain
services/deck and avalon services.

Replicate the secret-villain per-mode currentTurnState pattern for Avalon
and Clocktower: add avalon/utils/turn-state.ts and clocktower/utils/turn-state.ts
(each returning its own mode's TurnState), export from the mode barrel, and
import in every action file plus services.ts. Remove the local copies.
