+++ src/lib/game/modes/avalon/actions/propose-team.ts
@@ -1,13 +1,7 @@
 import type { Game, GameAction } from "@/lib/types";
-import { GameStatus } from "@/lib/types";
 
-import type { AvalonTurnState } from "../types";
 import { AvalonPhase } from "../types";
-
-function currentTurnState(game: Game): AvalonTurnState | undefined {
-  if (game.status.type !== GameStatus.Playing) return undefined;
-  return game.status.turnState as AvalonTurnState | undefined;
-}
+import { currentTurnState } from "../utils";
 
 export const proposeTeamAction: GameAction = {
   isValid(game: Game, callerId: string, payload: unknown) {
