+++ src/lib/game/modes/avalon/actions/resolve-quest.ts
@@ -2,13 +2,8 @@ import type { Game, GameAction } from "@/lib/types";
 import { GameStatus } from "@/lib/types";
 
 import { AvalonRole } from "../roles";
-import type { AvalonTurnState } from "../types";
 import { AvalonPhase, QuestCard } from "../types";
-
-function currentTurnState(game: Game): AvalonTurnState | undefined {
-  if (game.status.type !== GameStatus.Playing) return undefined;
-  return game.status.turnState as AvalonTurnState | undefined;
-}
+import { currentTurnState } from "../utils";
 
 const QUEST_WIN_THRESHOLD = 3;
