+++ src/lib/game/modes/avalon/actions/resolve-team-vote.ts
@@ -1,13 +1,8 @@
 import type { Game, GameAction } from "@/lib/types";
 import { GameStatus } from "@/lib/types";
 
-import type { AvalonTurnState } from "../types";
 import { AvalonPhase, TeamVote } from "../types";
-
-function currentTurnState(game: Game): AvalonTurnState | undefined {
-  if (game.status.type !== GameStatus.Playing) return undefined;
-  return game.status.turnState as AvalonTurnState | undefined;
-}
+import { currentTurnState } from "../utils";
 
 const CONSECUTIVE_REJECTION_LIMIT = 5;
