+++ src/lib/game/modes/avalon/index.ts
@@ -4,3 +4,4 @@ export * from "./lobby-config";
 export * from "./player-state";
 export * from "./roles";
 export * from "./types";
+export * from "./utils";
