+++ src/lib/game/modes/clocktower/actions/cast-public-vote.ts
@@ -1,14 +1,8 @@
 import type { Game, GameAction } from "@/lib/types";
-import { GameStatus } from "@/lib/types";
 
 import { ClocktowerRole } from "../roles";
-import type { ClocktowerTurnState } from "../types";
 import { ClocktowerPhase } from "../types";
-
-function currentTurnState(game: Game): ClocktowerTurnState | undefined {
-  if (game.status.type !== GameStatus.Playing) return undefined;
-  return game.status.turnState as ClocktowerTurnState | undefined;
-}
+import { currentTurnState } from "../utils";
 
 function getPlayerRole(
   game: Game,
