+++ src/lib/game/modes/clocktower/actions/nominate-player.ts
@@ -1,5 +1,4 @@
 import type { Game, GameAction } from "@/lib/types";
-import { GameStatus } from "@/lib/types";
 
 import {
   CLOCKTOWER_ROLES,
@@ -8,11 +7,7 @@ import {
 } from "../roles";
 import type { ClocktowerTurnState } from "../types";
 import { ClocktowerPhase } from "../types";
-
-function currentTurnState(game: Game): ClocktowerTurnState | undefined {
-  if (game.status.type !== GameStatus.Playing) return undefined;
-  return game.status.turnState as ClocktowerTurnState | undefined;
-}
+import { currentTurnState } from "../utils";
 
 function getPlayerRole(
   game: Game,
