+++ src/lib/game/modes/clocktower/actions/provide-information.ts
@@ -1,14 +1,9 @@
 import type { Game, GameAction } from "@/lib/types";
-import { GameStatus } from "@/lib/types";
 
 import { isClocktowerRole } from "../roles";
-import type { ClocktowerNightInformation, ClocktowerTurnState } from "../types";
+import type { ClocktowerNightInformation } from "../types";
 import { ClocktowerPhase } from "../types";
-
-function currentTurnState(game: Game): ClocktowerTurnState | undefined {
-  if (game.status.type !== GameStatus.Playing) return undefined;
-  return game.status.turnState as ClocktowerTurnState | undefined;
-}
+import { currentTurnState } from "../utils";
 
 interface InfoCandidate {
   type?: unknown;
