+++ src/lib/game/modes/clocktower/index.ts
@@ -5,3 +5,4 @@ export * from "./player-state";
 export * from "./roles";
 export * from "./services";
 export * from "./types";
+export * from "./utils";
