+++ src/lib/game/modes/secret-villain/services.ts
@@ -1,3 +1,4 @@
+import { shuffle } from "@/lib/game/shuffle";
 import { resolvePlayerOrder } from "@/lib/player-order";
 import type { Game, GameModeServices, PlayerRoleAssignment } from "@/lib/types";
 import { GameStatus, isSecretVillainModeConfig, Team } from "@/lib/types";
@@ -23,17 +24,6 @@ import {
   SecretVillainWinner,
   SvVictoryConditionKey,
 } from "./utils";
-
-function shuffle<T>(array: T[]): T[] {
-  const result = [...array];
-  for (let i = result.length - 1; i > 0; i--) {
-    const j = Math.floor(Math.random() * (i + 1));
-    const temp = result[i] as T;
-    result[i] = result[j] as T;
-    result[j] = temp;
-  }
-  return result;
-}
 
 function buildPhaseInfo(
   phase: SecretVillainTurnState["phase"],
