+++ src/lib/game/shuffle.ts
@@ -1,0 +1,14 @@
+/**
+ * Fisher-Yates shuffle. Returns a new shuffled array without mutating the
+ * input.
+ */
+export function shuffle<T>(array: T[]): T[] {
+  const result = [...array];
+  for (let i = result.length - 1; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    const temp = result[i] as T;
+    result[i] = result[j] as T;
+    result[j] = temp;
+  }
+  return result;
+}
